==> notebooks/step1-scrape_and_clean_data.ipynb <==
# Scrape SubReddit Pages and Preliminary cleaning

import pandas as pd
import numpy as np
import requests
import json
import csv
import time
import datetime as dt
import pandas as pd
from pmaw import PushshiftAPI

## Define a function to scrape data from SubReddit pages

PMAW: Under the hood, pmaw makes several API requests to Pushshift that each return a maximum of 100 comments, with requests being subjected to a rate-limit of 60 requests per minute. 


https://medium.com/swlh/how-to-scrape-large-amounts-of-reddit-data-using-pushshift-1d33bde9286


api = PushshiftAPI()

# Function to scrape subreddit submissions (different than comments)

def get_reddit_submissions(subreddit, limit):
    subreddit = subreddit
    limit = limit
    posts = api.search_submissions(subreddit=subreddit, 
                                   limit=limit, 
                                   before=before, 
                                   after=after)
    post_list = [post for post in posts]
    df = pd.DataFrame(post_list)
    return df

## Scrape Data from Pre-COVID Shutdown

Using the Pushshift API, scrape reddit data from before Covid Shutdown for each subredit of interest (Using 3/20/2020 as that was the date New York and California shut down)

before = int(dt.datetime(2020,3,20,0,0).timestamp()) # converts to UTC # A chosen 'shutdown' date
after = int(dt.datetime(2017,3,20,0,0).timestamp()) # converts to UTC 

### r/mentalhealth

#### Scrape comments

mentalhealth_PreCovid = get_reddit_submissions('mentalhealth', 5000)

# Check the shape to make sure all rows are there
mentalhealth_PreCovid.shape

#### Save columns of interest

# Keep columns: 'author', 'body', 'created_utc', 'subreddit'
mentalhealth_PreCovid = mentalhealth_PreCovid[['author', 
                                               'title', 
                                               'selftext',
                                               'created_utc', 
                                               'subreddit']]

#### Combine title with self text column 

mentalhealth_PreCovid['text'] = mentalhealth_PreCovid[['title', 
                                                       'selftext']].stack().groupby(level=0).agg(' : '.join)

# Drop the columns we combined
mentalhealth_PreCovid.drop(columns = ['title','selftext'], inplace=True)

#### Remove any 'authors' or 'body' rows where it says removed or deleted

drop_rows2 = mentalhealth_PreCovid.loc[mentalhealth_PreCovid['author'] == '[deleted]'].index

# Replace 'deleted' and 'removed' with null and then re
mentalhealth_PreCovid.drop(drop_rows2, inplace = True)

#### Drop duplicates

mentalhealth_PreCovid = mentalhealth_PreCovid.drop_duplicates(subset='text', keep="first")

mentalhealth_PreCovid.shape

### r/depression

#### Scrape comments

depression_PreCovid = get_reddit_submissions('depression', 5000)

depression_PreCovid.shape

#### Save columns of interest

depression_PreCovid = depression_PreCovid[['author', 
                                           'title', 
                                           'selftext',
                                           'created_utc', 
                                           'subreddit']]

#### Combine title with self text column and then check NAs again

depression_PreCovid['text'] = depression_PreCovid[['title', 
                                                   'selftext']].stack().groupby(level=0).agg(' : '.join)

# Drop the columns we combined
depression_PreCovid.drop(columns = ['title','selftext'], inplace=True)

#### Remove any 'authors' or 'body' rows where it says removed or deleted

drop_rows = depression_PreCovid.loc[depression_PreCovid['author'] == '[deleted]'].index

# Replace 'deleted' and 'removed' with null and then re
depression_PreCovid.drop(drop_rows, inplace = True)

#### Drop duplicates

depression_PreCovid = depression_PreCovid.drop_duplicates(subset='text', keep="first")

depression_PreCovid.shape

### r/Anxiety

#### Scrape comments

anxiety_PreCovid = get_reddit_submissions('Anxiety', 5000)

anxiety_PreCovid.shape

#### Save columns of interest

# Keep columns: 'author', 'body', 'created_utc', 'subreddit'
anxiety_PreCovid = anxiety_PreCovid[['author', 
                                     'title', 
                                     'selftext',
                                     'created_utc', 
                                     'subreddit']]

#### Combine title with self text column

anxiety_PreCovid['text'] = anxiety_PreCovid[['title', 
                                            'selftext']].stack().groupby(level=0).agg(' : '.join)

# Drop the columns we combined
anxiety_PreCovid.drop(columns = ['title','selftext'], inplace=True)

#### Remove any 'authors' or 'body' rows where it says removed or deleted

drop_rows = anxiety_PreCovid.loc[anxiety_PreCovid['author'] == '[deleted]'].index

# Replace 'deleted' and 'removed' with null and then re
anxiety_PreCovid.drop(drop_rows, inplace = True)

#### Drop duplicates

anxiety_PreCovid = anxiety_PreCovid.drop_duplicates(subset='text', keep="first")

anxiety_PreCovid.shape

### r/bipolar

#### Scrape comments

bipolar_PreCovid = get_reddit_submissions('bipolar', 5000)

bipolar_PreCovid.shape

#### Save columns of interest

bipolar_PreCovid = bipolar_PreCovid[['author', 
                                     'title', 
                                     'selftext',
                                     'created_utc', 
                                     'subreddit']]

#### Combine title with self text column

bipolar_PreCovid['text'] = bipolar_PreCovid[['title', 
                                            'selftext']].stack().groupby(level=0).agg(' : '.join)

# Drop the columns we combined
bipolar_PreCovid.drop(columns = ['title','selftext'], inplace=True)

#### Remove any 'authors' or 'body' rows where it says removed or deleted

drop_rows = bipolar_PreCovid.loc[bipolar_PreCovid['author'] == '[deleted]'].index

# Replace 'deleted' and 'removed' with null and then re
bipolar_PreCovid.drop(drop_rows, inplace = True)

#### Drop duplicates

bipolar_PreCovid = bipolar_PreCovid.drop_duplicates(subset='text', keep="first")

bipolar_PreCovid.shape

### r/BPD

##### Scrape comments

bpd_PreCovid = get_reddit_submissions('BPD', 5000)

bpd_PreCovid.shape

##### Save columns of interest

bpd_PreCovid = bpd_PreCovid[['author', 
                             'title', 
                             'selftext',
                             'created_utc', 
                             'subreddit']]

#### Combine title with self text column

bpd_PreCovid['text'] = bpd_PreCovid[['title', 
                                            'selftext']].stack().groupby(level=0).agg(' : '.join)

# Drop the columns we combined
bpd_PreCovid.drop(columns = ['title','selftext'], inplace=True)

#### Remove any 'authors' or 'body' rows where it says removed or deleted

drop_rows = bpd_PreCovid.loc[bpd_PreCovid['author'] == '[deleted]'].index

# Replace 'deleted' and 'removed' with null and then re
bpd_PreCovid.drop(drop_rows, inplace = True)

#### Drop duplicates

bpd_PreCovid = bpd_PreCovid.drop_duplicates(subset='text', keep="first")

bpd_PreCovid.shape

### r/schizophrenia

schizophrenia_PreCovid = get_reddit_submissions('schizophrenia', 5000)

schizophrenia_PreCovid.shape

##### Save columns of interest

schizophrenia_PreCovid = schizophrenia_PreCovid[['author', 
                                                 'title', 
                                                 'selftext',
                                                 'created_utc', 
                                                 'subreddit']]

#### Combine title with self text column

schizophrenia_PreCovid['text'] = schizophrenia_PreCovid[['title', 
                                                         'selftext']].stack().groupby(level=0).agg(' : '.join)

# Drop the columns we combined
schizophrenia_PreCovid.drop(columns = ['title','selftext'], inplace=True)

#### Remove any 'authors' or 'body' rows where it says removed or deleted

drop_rows = schizophrenia_PreCovid.loc[schizophrenia_PreCovid['author'] == '[deleted]'].index

# Replace 'deleted' and 'removed' with null and then re
schizophrenia_PreCovid.drop(drop_rows, inplace = True)

#### Drop duplicates

schizophrenia_PreCovid = schizophrenia_PreCovid.drop_duplicates(subset='text', keep="first")

schizophrenia_PreCovid.shape

### r/autism

autism_PreCovid = get_reddit_submissions('autism', 5000)

autism_PreCovid.shape

##### Save columns of interest

autism_PreCovid = autism_PreCovid[['author', 
                                   'title', 
                                   'selftext',
                                   'created_utc', 
                                   'subreddit']]

#### Combine title with self text column

autism_PreCovid['text'] = autism_PreCovid[['title', 
                                           'selftext']].stack().groupby(level=0).agg(' : '.join)

# Drop the columns we combined
autism_PreCovid.drop(columns = ['title','selftext'], inplace=True)

#### Remove any 'authors' or 'body' rows where it says removed or deleted

drop_rows = autism_PreCovid.loc[autism_PreCovid['author'] == '[deleted]'].index

# Replace 'deleted' and 'removed' with null and then re
autism_PreCovid.drop(drop_rows, inplace = True)

#### Drop duplicates

autism_PreCovid = autism_PreCovid.drop_duplicates(subset='text', keep="first")

autism_PreCovid.shape

### r/AnorexiaNervosa

anorexia_PreCovid = get_reddit_submissions('AnorexiaNervosa', 5000)

anorexia_PreCovid.shape

##### Save columns of interest

anorexia_PreCovid = anorexia_PreCovid[['author', 
                                       'title', 
                                       'selftext',
                                       'created_utc', 
                                       'subreddit']]

#### Combine title with self text column

anorexia_PreCovid['text'] = anorexia_PreCovid[['title', 
                                               'selftext']].stack().groupby(level=0).agg(' : '.join)

# Drop the columns we combined
anorexia_PreCovid.drop(columns = ['title','selftext'], inplace=True)

#### Remove any 'authors' or 'body' rows where it says removed or deleted

drop_rows = anorexia_PreCovid.loc[anorexia_PreCovid['author'] == '[deleted]'].index

# Replace 'deleted' and 'removed' with null and then re
anorexia_PreCovid.drop(drop_rows, inplace = True)

#### Drop duplicates

anorexia_PreCovid = anorexia_PreCovid.drop_duplicates(subset='text', keep="first")

anorexia_PreCovid.shape

### r/Bulimia

bulimia_PreCovid = get_reddit_submissions('Bulimia', 5000)

bulimia_PreCovid.shape

##### Save columns of interest

bulimia_PreCovid = bulimia_PreCovid[['author', 
                                     'title', 
                                     'selftext',
                                     'created_utc', 
                                     'subreddit']]

#### Combine title with self text column

bulimia_PreCovid['text'] = bulimia_PreCovid[['title', 
                                             'selftext']].stack().groupby(level=0).agg(' : '.join)

# Drop the columns we combined
bulimia_PreCovid.drop(columns = ['title','selftext'], inplace=True)

#### Remove any 'authors' or 'body' rows where it says removed or deleted

drop_rows = bulimia_PreCovid.loc[bulimia_PreCovid['author'] == '[deleted]'].index

# Replace 'deleted' and 'removed' with null and then re
bulimia_PreCovid.drop(drop_rows, inplace = True)

#### Drop duplicates

bulimia_PreCovid = bulimia_PreCovid.drop_duplicates(subset='text', keep="first")

bulimia_PreCovid.shape

### Combine data

#### Combine data to one large data frame

precovid_frames = [bulimia_PreCovid, 
                   anorexia_PreCovid,
                   autism_PreCovid,
                   schizophrenia_PreCovid,
                   bpd_PreCovid,
                   bipolar_PreCovid,
                   anxiety_PreCovid,
                   depression_PreCovid,
                   mentalhealth_PreCovid]

PreCovid_subreddits = pd.concat(precovid_frames)

PreCovid_subreddits.shape

## Scrape Data from Post-COVID Shutdown

before = int(dt.datetime(2021,3,12,0,0).timestamp()) # converts to UTC 
after = int(dt.datetime(2020,3,20,0,0).timestamp()) # converts to UTC # A chosen 'shutdown' date

### r/mentalhealth

mentalhealth_PostCovid = get_reddit_submissions('mentalhealth', 5000)

mentalhealth_PostCovid.shape 

##### Save columns of interest

mentalhealth_PostCovid = mentalhealth_PostCovid[['author', 
                                                 'title', 
                                                 'selftext', 
                                                 'created_utc', 
                                                 'subreddit']]

#### Combine title with self text column

mentalhealth_PostCovid['text'] = mentalhealth_PostCovid[['title', 
                                                         'selftext']].stack().groupby(level=0).agg(' : '.join)

# Drop the columns we combined
mentalhealth_PostCovid.drop(columns = ['title','selftext'], inplace=True)

#### Remove any 'authors' or 'body' rows where it says removed or deleted

drop_rows = mentalhealth_PostCovid.loc[mentalhealth_PostCovid['author'] == '[deleted]'].index

# Replace 'deleted' and 'removed' with null and then re
mentalhealth_PostCovid.drop(drop_rows, inplace = True)

#### Drop duplicates

mentalhealth_PostCovid = mentalhealth_PostCovid.drop_duplicates(subset='text', keep="first")

mentalhealth_PostCovid.shape

### r/depression

depression_PostCovid = get_reddit_submissions('depression', 5000)

depression_PostCovid.shape

##### Save columns of interest

depression_PostCovid = depression_PostCovid[['author', 
                                             'title', 
                                             'selftext', 
                                             'created_utc', 
                                             'subreddit']]

#### Combine title with self text column

depression_PostCovid['text'] = depression_PostCovid[['title', 
                                                     'selftext']].stack().groupby(level=0).agg(' : '.join)

# Drop the columns we combined
depression_PostCovid.drop(columns = ['title','selftext'], inplace=True)

#### Remove any 'authors' or 'body' rows where it says removed or deleted

drop_rows = depression_PostCovid.loc[depression_PostCovid['author'] == '[deleted]'].index

# Replace 'deleted' and 'removed' with null and then re
depression_PostCovid.drop(drop_rows, inplace = True)

#### Drop duplicates

depression_PostCovid = depression_PostCovid.drop_duplicates(subset='text', keep="first")

depression_PostCovid.shape

### r/Anxiety

anxiety_PostCovid = get_reddit_submissions('Anxiety', 5000)

anxiety_PostCovid.shape 

##### Save columns of interest

anxiety_PostCovid = anxiety_PostCovid[['author', 
                                       'title', 
                                       'selftext', 
                                       'created_utc', 
                                       'subreddit']]

#### Combine title with self text column

anxiety_PostCovid['text'] = anxiety_PostCovid[['title', 
                                               'selftext']].stack().groupby(level=0).agg(' : '.join)

# Drop the columns we combined
anxiety_PostCovid.drop(columns = ['title','selftext'], inplace=True)

#### Remove any 'authors' or 'body' rows where it says removed or deleted

drop_rows = anxiety_PostCovid.loc[anxiety_PostCovid['author'] == '[deleted]'].index

# Replace 'deleted' and 'removed' with null and then re
anxiety_PostCovid.drop(drop_rows, inplace = True)

#### Drop duplicates

anxiety_PostCovid = anxiety_PostCovid.drop_duplicates(subset='text', keep="first")

anxiety_PostCovid.shape

### r/bipolar

bipolar_PostCovid = get_reddit_submissions('bipolar', 5000)

bipolar_PostCovid.shape

##### Save columns of interest

bipolar_PostCovid = bipolar_PostCovid[['author', 
                                       'title', 
                                       'selftext',
                                       'created_utc', 
                                       'subreddit']]

#### Combine title with self text column

bipolar_PostCovid['text'] = bipolar_PostCovid[['title', 
                                               'selftext']].stack().groupby(level=0).agg(' : '.join)

# Drop the columns we combined
bipolar_PostCovid.drop(columns = ['title','selftext'], inplace=True)

#### Remove any 'authors' or 'body' rows where it says removed or deleted

drop_rows = bipolar_PostCovid.loc[bipolar_PostCovid['author'] == '[deleted]'].index

# Replace 'deleted' and 'removed' with null and then re
bipolar_PostCovid.drop(drop_rows, inplace = True)

#### Drop duplicates

bipolar_PostCovid = bipolar_PostCovid.drop_duplicates(subset='text', keep="first")

bipolar_PostCovid.shape

### r/BPD

bpd_PostCovid = get_reddit_submissions('BPD', 5000)

bpd_PostCovid.shape

##### Save columns of interest

bpd_PostCovid = bpd_PostCovid[['author',
                               'title', 
                               'selftext',
                               'created_utc', 
                               'subreddit']]

#### Combine title with self text column

bpd_PostCovid['text'] = bpd_PostCovid[['title','selftext']].stack().groupby(level=0).agg(' : '.join)

# Drop the columns we combined
bpd_PostCovid.drop(columns = ['title','selftext'], inplace=True)

#### Remove any 'authors' or 'body' rows where it says removed or deleted

drop_rows = bpd_PostCovid.loc[bpd_PostCovid['author'] == '[deleted]'].index

# Replace 'deleted' and 'removed' with null and then re
bpd_PostCovid.drop(drop_rows, inplace = True)

#### Drop duplicates

bpd_PostCovid = bpd_PostCovid.drop_duplicates(subset='text', keep="first")

bpd_PostCovid.shape

### r/schizophrenia

schizophrenia_PostCovid = get_reddit_submissions('schizophrenia', 5000)

schizophrenia_PostCovid.shape 

##### Save columns of interest

schizophrenia_PostCovid = schizophrenia_PostCovid[['author', 
                                                   'title', 
                                                   'selftext',
                                                   'created_utc',
                                                   'subreddit']]

#### Combine title with self text column

schizophrenia_PostCovid['text'] = schizophrenia_PostCovid[['title',
                                                           'selftext']].stack().groupby(level=0).agg(' : '.join)

# Drop the columns we combined
schizophrenia_PostCovid.drop(columns = ['title','selftext'], inplace=True)

#### Remove any 'authors' or 'body' rows where it says removed or deleted

drop_rows = schizophrenia_PostCovid.loc[schizophrenia_PostCovid['author'] == '[deleted]'].index

# Replace 'deleted' and 'removed' with null and then re
schizophrenia_PostCovid.drop(drop_rows, inplace = True)

#### Drop duplicates

schizophrenia_PostCovid = schizophrenia_PostCovid.drop_duplicates(subset='text', keep="first")

schizophrenia_PostCovid.shape

### r/autism

autism_PostCovid = get_reddit_submissions('autism', 5000)

autism_PostCovid.shape

##### Save columns of interest

autism_PostCovid = autism_PostCovid[['author', 
                                     'title', 
                                     'selftext',
                                     'created_utc', 
                                     'subreddit']]

#### Combine title with self text column

autism_PostCovid['text'] = autism_PostCovid[['title','selftext']].stack().groupby(level=0).agg(' : '.join)

# Drop the columns we combined
autism_PostCovid.drop(columns = ['title','selftext'], inplace=True)

#### Remove any 'authors' or 'body' rows where it says removed or deleted

drop_rows = autism_PostCovid.loc[autism_PostCovid['author'] == '[deleted]'].index

# Replace 'deleted' and 'removed' with null and then re
autism_PostCovid.drop(drop_rows, inplace = True)

#### Drop duplicates

autism_PostCovid  = autism_PostCovid.drop_duplicates(subset='text', keep="first")

autism_PostCovid .shape

### r/AnorexiaNervosa

anorexia_PostCovid = get_reddit_submissions('AnorexiaNervosa', 5000)

anorexia_PostCovid.shape

##### Save columns of interest

anorexia_PostCovid = anorexia_PostCovid[['author', 
                                         'title', 
                                         'selftext',
                                         'created_utc', 
                                         'subreddit']]

#### Combine title with self text column

anorexia_PostCovid['text'] = anorexia_PostCovid[['title','selftext']].stack().groupby(level=0).agg(' : '.join)

# Drop the columns we combined
anorexia_PostCovid.drop(columns = ['title','selftext'], inplace=True)

#### Remove any 'authors' or 'body' rows where it says removed or deleted

drop_rows = anorexia_PostCovid.loc[anorexia_PostCovid['author'] == '[deleted]'].index

# Replace 'deleted' and 'removed' with null and then re
anorexia_PostCovid.drop(drop_rows, inplace = True)

#### Drop duplicates

anorexia_PostCovid = anorexia_PostCovid.drop_duplicates(subset='text', keep="first")

anorexia_PostCovid.shape

### r/Bulimia

bulimia_PostCovid = get_reddit_submissions('Bulimia', 5000)

bulimia_PostCovid.shape

##### Save columns of interest

bulimia_PostCovid = bulimia_PostCovid[['author', 
                                       'title', 
                                       'selftext',
                                       'created_utc', 
                                       'subreddit']]

#### Combine title with self text column

bulimia_PostCovid['text'] = bulimia_PostCovid[['title','selftext']].stack().groupby(level=0).agg(' : '.join)

# Drop the columns we combined
bulimia_PostCovid.drop(columns = ['title','selftext'], inplace=True)

#### Remove any 'authors' or 'body' rows where it says removed or deleted

drop_rows = bulimia_PostCovid.loc[bulimia_PostCovid['author'] == '[deleted]'].index

# Replace 'deleted' and 'removed' with null and then re
bulimia_PostCovid.drop(drop_rows, inplace = True)

#### Drop duplicates

bulimia_PostCovid = bulimia_PostCovid.drop_duplicates(subset='text', keep="first")

bulimia_PostCovid.shape

### Combine data

#### Combine data to one large data frame

postcovid_frames = [bulimia_PostCovid, 
                   anorexia_PostCovid,
                   autism_PostCovid,
                   schizophrenia_PostCovid,
                   bpd_PostCovid,
                   bipolar_PostCovid,
                   anxiety_PostCovid,
                   depression_PostCovid,
                   mentalhealth_PostCovid]

PostCovid_subreddits = pd.concat(postcovid_frames)

#### Combine pre and post Covid data frames

# add column to each data frame saying 'pre or post covid'
PreCovid_subreddits['timeframe'] = 'pre-covid'
PostCovid_subreddits['timeframe'] = 'post-covid'

# Combine data frames
subreddit_posts_combined = pd.concat([PreCovid_subreddits, PostCovid_subreddits], axis=0)

#### Anonmyize


# Replace unique author id with integer.
subreddit_posts_combined['author'] = subreddit_posts_combined.assign(author=(np.unique(subreddit_posts_combined.author, 
                                          return_inverse=True)[1] + 1).astype(str))

# Add 'sub' infront of integer to remind ourselves that this is an id. 
subreddit_posts_combined['author'] = subreddit_posts_combined['author'].apply(lambda x: f"sub{x}")

#### Save to csv

# Save combined data
subreddit_posts_combined.to_csv('../data/subreddit_posts_combined.csv',
                                index=False,
                                columns=list(subreddit_posts_combined.axes[1]))
