==> notebooks/step2-pre-process_text.ipynb <==
# Text Pre-processing

Add text here explaining this notebook and what it does

##### Import libraries

import pandas as pd
import numpy as np

import seaborn as sns
import matplotlib.pyplot as plt
plt.style.use('fivethirtyeight')

import re
import string

import nltk
from nltk import word_tokenize, sent_tokenize
from nltk.stem import PorterStemmer
from sklearn.feature_extraction.text import TfidfVectorizer
from nltk.corpus import stopwords

## Load data

# Read in combined data frames csv file
posts = pd.read_csv('../data/subreddit_posts_combined.csv') 

## Look at the data

posts.head()

#### Check Shape

posts.shape

#### Check data types

posts.dtypes

#### Check Null Values

posts.isnull().sum()

(posts.loc[posts['text'].isnull() == True]).head(10)

posts.dropna(inplace=True)

posts.shape

#### Count number of posts per subreddit

posts['subreddit'].value_counts()

## Preprocess Text

### Normalization

For the stop words, I'm not sure it makes sense to include words like anxious, anxiety, depressed, and depression as these could be included in any of these subreddit posts as it's a feeling in addition to a disorder. We will include as stop words for now to see which words we find for those disorders that are not the title of the subreddit.

# For removing stopwords
stop_words = set(stopwords.words('english'))  

new_stopwords = ['anorexia', 'nervosa', 'anorexic', 'bulimic', 'bulimia', 'bipolar',
                 'anxiety', 'anxious','autistic', 'autism', 'bpd', 'borderline',
                 'schizophrenia', 'schizophrenic', 'schizo', 'depression', 
                 'depressed']

stop_words.update(new_stopwords)
new_stopwords_list = set(stop_words)



# To only use words within the NLTK corpus, and remove nonsense words
words = set(nltk.corpus.words.words())

https://towardsdatascience.com/cleaning-text-data-with-python-b69b47b97b76

def text_preproc(x):
    x = x.lower()
    x = x.encode('ascii', 'ignore').decode("utf-8")
    x = re.sub(r'https*\S+', ' ', x)
    x = re.sub(r'@\S+', ' ', x)
    x = re.sub(r'#\S+', ' ', x)
    x = re.sub(r'\'\w+', '', x)
    x = re.sub('[%s]' % re.escape(string.punctuation), ' ', x)
    x = re.sub(r'\w*\d+\w*', '', x)
    x = re.sub(r'\s{2,}', ' ', x)
    x = " ".join(w for w in nltk.wordpunct_tokenize(x) if w.lower() \
                 not in new_stopwords_list)
    x = " ".join(w for w in nltk.wordpunct_tokenize(x) if w.lower() \
                 in words or not w.isalpha())
    return x

posts['text_clean'] = posts.text.apply(text_preproc)

posts.head()

### Tokenize sentences and words

# tokenize sentences
posts['sent_tokens'] = [sent_tokenize(row) for row in posts['text_clean'].astype(str)] 

# tokenize words
posts['word_tokens'] = [word_tokenize(row) for row in posts['text_clean'].astype(str)]

### Check for empty cells

posts.replace('', np.nan, inplace=True) 

posts.isnull().sum()

posts.dropna(inplace=True)

### Porter Stemmer

Using Porter Stemmer

# Create an instance of Porter Stemmer
stemmer = PorterStemmer()

text = posts['word_tokens'].astype(str)

posts['word_tokens'] = [stemmer.stem(word) for word in text]

### More cleaning

### Check dataframe filtered by submission Lengths

posts.head()

# Create a new column with submission length
posts['text_length'] = posts['sent_tokens'].map(len)

# Create a new column with number of words per submission
posts['stems_word_count'] = posts['word_tokens'].map(len)

# Show the shortest submissions
posts.sort_values(by='text_length', ascending=True)[['sent_tokens']].head()

# Show the shortest submissions
posts.sort_values(by='stems_word_count', ascending=True)[['word_tokens']].head()

# Show the longest submissions
posts.sort_values(by='stems_word_count', ascending=False)[['word_tokens']].head(10)

# Drop any unecessary rows before saving the dataframe to a clean csv file
posts.drop(columns=['text','text_length','stems_word_count'], inplace=True)

## Save to csv

posts.to_csv(r'../data/posts-preprocessed.csv', index=False)
