==> notebooks/step5-word2vec.ipynb <==
# Word2Vec 

##### Import libraries

import pandas as pd

import numpy as np
from numpy import array
from numpy import zeros

import matplotlib.pyplot as plt

import re
import multiprocessing

# Word2Vec
import gensim 
import gensim.models
from gensim.models import Word2Vec
from gensim.test.utils import datapath
from gensim import utils

##### Load data

posts = pd.read_csv('../data/posts-preprocessed.csv')

posts.head()

##  Create Word2Vec model

Using **Continuous Bag of Words (CBOW) model**: CBOW model predicts the current word given context words within specific window. The input layer contains the context words and the output layer contains the current word. The hidden layer contains the number of dimensions in which we want to represent current word present at the output layer.

https://www.geeksforgeeks.org/python-word-embedding-using-word2vec/

# How many cores am I working with?
cores = multiprocessing.cpu_count() # Count the number of cores in this computer
cores

Parameters: 
- size: (default 100) The number of dimensions of the embedding, e.g. the length of the dense vector to represent each token (word).
- window: (default 5) The maximum distance between a target word and words around the target word.
- min_count: (default 5) The minimum count of words to consider when training the model; words with an occurrence less than this count will be ignored.
- workers: (default 3) The number of threads to use while training.
- sg: (default 0 or CBOW) The training algorithm, either CBOW (0) or skip gram (1).

Code help from:
- https://radimrehurek.com/gensim/auto_examples/tutorials/run_word2vec.html
- https://machinelearningmastery.com/develop-word-embedding-model-predicting-movie-review-sentiment/

The word2vec algorithm processes documents sentence by sentence. This means we will preserve the sentence-based structure during cleaning.

posts.head(1)

### Training a word2vec model on reddit posts

Preprocessing of sentences using Gensim's pre-processing on posts, so that the input yields one sentence (list of utf8 words) after another. https://radimrehurek.com/gensim/auto_examples/tutorials/run_word2vec.html

class MyCorpus:
    """An iterator that yields sentences (lists of str)."""

    def __iter__(self):
        for line in posts['sent_tokens']:
            # assume there's one document per line, tokens separated by whitespace
            yield utils.simple_preprocess(line)

sentences = MyCorpus()

model = gensim.models.Word2Vec(sentences=sentences, 
                               window = 10,
                               size = 150,
                               min_count = 2,
                               workers=cores-1) # run on all cores minus 1

After the model is fit, we print the size of the learned vocabulary

# summarize vocabulary size in model
vocab = list(model.wv.vocab)
print('Vocabulary size: %d' % len(vocab))

Finally, we save the learned embedding vectors to file using the save_word2vec_format() on the model’s ‘wv‘ (word vector) attribute. The embedding is saved in ASCII format with one word and vector per line.

filename = '../embedding_word2vec.txt'

# save model in ASCII (word2vec) format
model.wv.save_word2vec_format(filename, binary=False)
